Update resolvers, add command resolver
